% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rao_index.R
\name{get_rao_index}
\alias{get_rao_index}
\title{Rao's index of a numeric vector}
\usage{
get_rao_index(numeric_v)
}
\arguments{
\item{numeric_v}{numeric vector of 3 values at the very least}
}
\value{
a scalar value (the rao's index)
}
\description{
Calculates Rao's index.
}
\details{
Rao's quadratic entropy is a measure of diversity of ecological communities
defined by Rao (1982) and is based on the proportion of the abundance of
species present in a community and some measure of distance (dissimilarity)
among them. Dissimilarity is defined as the euclidean distance.
\deqn{\rho = \deqn{\sum_{ij}d_{ij} p_i p_j }}

The function accepts a numeric value vector (or matrix of numeric values) and returns Rao's index.
It is particularly useful in remote sensing applications as it can be used to calculate a temporal diversity index
between layers of raster stacks.
}
\examples{
i <- 1; j <- 200
m <- c(i, i, j, i, j, j, j, j, j)
get_rao_index(m)

i <- 201; j <- 200
m <- c(i, i, j, i, j, j, j, j, j)
get_rao_index(m)

# pass a matrix to get_rao_index, as a matrix is just a vector with an x and y dimensions
M <- matrix(m, byrow = TRUE, nrow = 3)
identical(get_rao_index(M), get_rao_index(m))
# [1] TRUE
}
\references{
Rao, C. R. (1982). Diversity and dissimilarity coefficients: a unified approach. Theoretical Population Biology, 21(1), 24–43
\url{https://doi.org/10.1016/0040-5809(82)90004-1}
}
