% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_maker.R
\name{scale_maker}
\alias{scale_maker}
\title{Scale a numeric vector by a product factor and round the result}
\usage{
scale_maker(prod_factor, digits = 1)
}
\arguments{
\item{prod_factor}{numeric vector of scaling factors}

\item{digits}{integer indicating the number of decimal places}
}
\value{
a function to scale by \code{prod_factor} and round to the \code{digits} digits. The function accepts a numeric vector
}
\description{
Scale a numeric vector by a product factor and round the result
}
\examples{
scale_by_ten <- scale_maker(10)
scale_by_ten(10)
# [1] 100
scale_by_ten(0.0046739)
# [1] 0
scale_by_ten <- scale_maker(10, 2)
scale_by_ten(0.0046739)
# [1] 0.05
}
