% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_to_nd.R
\name{scale_to_nd}
\alias{scale_to_nd}
\title{Scale raster layer to [-1, 1] range}
\usage{
scale_to_nd(raster_stack)
}
\arguments{
\item{raster_stack}{a Raster* object (RasterLayer, RasterStack, RasterBrick)}
}
\value{
a Radster* object scaled to [-1, 1] values
}
\description{
This is a helper function to scales values of a raster to the same range of
normalized difference indexes (NDVI, EVI, ecc). "Normalizing" rasters may be useful to enable
an optimal confrontation of Rao'index values resulting from get_rao_index function
among heterogenous rasters.
}
